Run code through 'futurize'

+++ src/gdata/Crypto/PublicKey/_slowmath.py
@@ -23,6 +23,7 @@
 # ===================================================================
 
 """Pure Python implementation of the RSA-related portions of Crypto.PublicKey._fastmath."""
+from future.builtins import object
 
 __revision__ = "$Id$"
 
@@ -81,12 +82,12 @@ class _RSAKey(object):
 
 def rsa_construct(n, e, d=None, p=None, q=None, u=None):
     """Construct an RSAKey object"""
-    assert isinstance(n, long)
-    assert isinstance(e, long)
-    assert isinstance(d, (long, type(None)))
-    assert isinstance(p, (long, type(None)))
-    assert isinstance(q, (long, type(None)))
-    assert isinstance(u, (long, type(None)))
+    assert isinstance(n, int)
+    assert isinstance(e, int)
+    assert isinstance(d, (int, type(None)))
+    assert isinstance(p, (int, type(None)))
+    assert isinstance(q, (int, type(None)))
+    assert isinstance(u, (int, type(None)))
     obj = _RSAKey()
     obj.n = n
     obj.e = e
@@ -151,7 +152,7 @@ class _DSAKey(object):
         # SECURITY TODO - We _should_ be computing SHA1(m), but we don't because that's the API.
         if not self.has_private():
             raise TypeError("No private key")
-        if not (1L < k < self.q):
+        if not (1 < k < self.q):
             raise ValueError("k is not between 2 and q-1")
         inv_k = inverse(k, self.q)   # Compute k**-1 mod q
         r = pow(self.g, k, self.p) % self.q  # r = (g**k mod p) mod q
@@ -169,11 +170,11 @@ class _DSAKey(object):
         return v == r
 
 def dsa_construct(y, g, p, q, x=None):
-    assert isinstance(y, long)
-    assert isinstance(g, long)
-    assert isinstance(p, long)
-    assert isinstance(q, long)
-    assert isinstance(x, (long, type(None)))
+    assert isinstance(y, int)
+    assert isinstance(g, int)
+    assert isinstance(p, int)
+    assert isinstance(q, int)
+    assert isinstance(x, (int, type(None)))
     obj = _DSAKey()
     obj.y = y
     obj.g = g
